Add UserCheckNamelyField decorator Remove duplication from Jobvite::Connection and Greenhouse::Connection
